+++ src/scripts/main/constants/keyboardKeys.constants.ts
@@ -11,3 +11,14 @@ export enum KeyboardKeys {
   Alt = 'Alt',
   Command = 'Meta',
 }
+
+export enum KeyCodes {
+  OpenInfoModal = 191,
+  QuitGame = 81,
+  SaveGame = 83,
+  DescendLevel = 190,
+  AscendLevel = 188,
+  OpenDevFeaturesModal = 84,
+  Examine = 88,
+  Activate = 65,
+}